App: split configuration from report assembly
Structure-only commit: app/config.rs holds the CLI surface — Args
(with its clap attributes), Config, CursorConfig, the default
log-location probes, cursor resolution, and the demo switch — i.e.
what the app reads and which mode it runs in. app.rs keeps run,
report loading, collector orchestration, and provider ordering.
Callers keep addressing app::{Args, Config}.

Second-model review round 1: one P3 (module-doc wording), fixed;
same-instant A/B against a HEAD worktree build renders
byte-identical.
